+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Show public/private event in Event detail panel [#622](https://github.com/rokwire/illinois-app/issues/622).
+- Pop to Group Detail Panel when user replies on a post [#623](https://github.com/rokwire/illinois-app/issues/623).
 
 ## [2.6.12] - 2021-07-27
 ### Fixed